Provide a menu to choose which cyoa data file to play.
Read the filesystem looking for cyoa_ game files in the examples/configuration directory.
Pulls in the files that match, strips off the .py extension and displays them to the user.
User uses either slider up/down to pick the file, and the screen button to choose.
Game then launches with selected data file.